🛠 GET -- /api/projects/:id/actions + middleware moved to own 📂
